add-admin-region-editing: cleaned up some things into its own file
add-admin-region-editing: cleaned up some things into its own file
