+++ MountainTracker.Client/src/app/graphql-helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './graphql-decorators'
 export * from './graphql-helper'
 export * from './ql-selection-set'
+export * from './graphql-meta-query'
