+++ MountainTracker.Client/src/app/services/graphql/local-service/base-ql.service.ts
@@ -1,59 +1,7 @@
 import { Apollo, ApolloBase, TypedDocumentNode, gql } from "apollo-angular";
-import { QlSelectionSet, QlSelectionSetTyped } from "../../../graphql-helpers";
-import * as graphqlHelpers from '../../../graphql-helpers/graphql-helper';
-import { Observable, map, of, switchMap } from "rxjs";
+import { QlQueryMeta, QlQueryParams, QlSelectionSet, QlSelectionSetTyped } from "../../../graphql-helpers";
+import { Observable, of, switchMap } from "rxjs";
 import { ApolloQueryResult } from "@apollo/client/core";
-
-type CoreQlNullableTypes =    'String' |  'Int' |  'Float' |  'Boolean' |  'ID'
-type CoreQlNonNullableTypes = 'String!' | 'Int!' | 'Float!' | 'Boolean!' | 'ID!'
-type CoreQlNullableListTypes =    '[String]' |  '[Int]' |  '[Float]' |  '[Boolean]' |  '[ID]'
-type CoreQlNonNullableListTypes = '[String]!' | '[Int]!' | '[Float]!' | '[Boolean]!' | '[ID]!'
-type NetQlNullableTypes =    'BigInt' |  'Byte' |  'Date' |  'DateOnly' |  'DateTime' |  'DateTimeOffset' |  'Decimal' |  'Guid' |  'Half' |  'Long' |  'Milliseconds' |  'SByte' |  'Seconds' |  'Short' |  'TimeOnly' |  'UInt' |  'ULong' |  'Uri' |  'UShort'
-type NetQlNonNullableTypes = 'BigInt!' | 'Byte!' | 'Date!' | 'DateOnly!' | 'DateTime!' | 'DateTimeOffset!' | 'Decimal!' | 'Guid!' | 'Half!' | 'Long!' | 'Milliseconds!' | 'SByte!' | 'Seconds!' | 'Short!' | 'TimeOnly!' | 'UInt!' | 'ULong!' | 'Uri!' | 'UShort!'
-type NetQlNullableListTypes =    '[BigInt]' |  '[Byte]' |  '[Date]' |  '[DateOnly]' |  '[DateTime]' |  '[DateTimeOffset]' |  '[Decimal]' |  '[Guid]' |  '[Half]' |  '[Long]' |  '[Milliseconds]' |  '[SByte]' |  '[Seconds]' |  '[Short]' |  '[TimeOnly]' |  '[UInt]' |  '[ULong]' |  '[Uri]' |  '[UShort]'
-type NetQlNonNullableListTypes = '[BigInt]!' | '[Byte]!' | '[Date]!' | '[DateOnly]!' | '[DateTime]!' | '[DateTimeOffset]!' | '[Decimal]!' | '[Guid]!' | '[Half]!' | '[Long]!' | '[Milliseconds]!' | '[SByte]!' | '[Seconds]!' | '[Short]!' | '[TimeOnly]!' | '[UInt]!' | '[ULong]!' | '[Uri]!' | '[UShort]!'
-type BasicQLTypes = CoreQlNullableTypes | CoreQlNonNullableTypes | CoreQlNullableListTypes | CoreQlNonNullableListTypes | NetQlNullableTypes | NetQlNonNullableTypes | NetQlNullableListTypes | NetQlNonNullableListTypes
-
-export type QlQueryParams = {
-  param: string, 
-  input?: string, 
-  qlType: BasicQLTypes | string
-};
-
-export class QlQueryMeta<T>{
-  public readonly queryParamPrefix: string;
-  public readonly queryParams: QlQueryParams[];
-  public readonly selection: string;
-  public readonly query: string;
-
-  public hasParamValues(): boolean
-  {
-    return this.queryParams && this.queryParams.length > 0;
-  }
-
-  public queryParamsFlat(): string
-  {
-    return this.queryParams.map((x)=>`$${this.queryParamPrefix}_${x.param}: ${x.qlType}`).join(',')
-  }
-
-  public getParamSelector(param:string): string
-  {
-    const index = this.queryParams.findIndex(x=>x.param === param)
-    return `${this.queryParamPrefix}_${this.queryParams[index].param}`
-  }
-
-  public constructor(classToField: T, query:string, selection?: QlSelectionSet | QlSelectionSetTyped<undefined, any>, queryParams: QlQueryParams[] = [], queryParamPrefix: string | null = null){
-    this.query = query;
-    this.queryParamPrefix = queryParamPrefix ?? (classToField as any).name
-    this.queryParams = queryParams;
-    const hasValues: boolean = this.hasParamValues();
-    const braceValues: (char: string)=>string = (char: string) => hasValues ? char : '';
-    this.selection = `${query}${braceValues('(')}${this.queryParams.map((x)=>`${x?.input ?? x.param}: $${this.queryParamPrefix}_${x.param}`).join(',')}${braceValues(')')}
-{
-${graphqlHelpers.selectToQlFields(selection) ?? graphqlHelpers.ensureQlFields((classToField as any).prototype.constructor)}
-}`
-  }
-}  
 
 export abstract class BaseQlService {
     public readonly moutainTrackerApi: ApolloBase;
