+++ MountainTracker.Client/src/app/services/graphql/local-service/region.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable, map, switchMap } from 'rxjs';
-import { QlSelectionSet, QlSelectionSetTyped } from '../../../graphql-helpers';
-import { BaseQlService, QlQueryMeta, QlQueryParams } from './base-ql.service';
+import { QlQueryMeta, QlQueryParams, QlSelectionSet, QlSelectionSetTyped } from '../../../graphql-helpers';
+import { BaseQlService } from './base-ql.service';
 import { Apollo } from 'apollo-angular';
 import { Region } from '../../../models';
